pyproject.toml
[project]
name = "rezehor-core"
version = "0.1.0"
description = "Personal AI Assistant"
authors = [{ name = "Timur", email = "timurkhvat@gmail.com" }]
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "anthropic>=0.76.0",
    "keyboard>=0.13.5",
    "loguru>=0.7.3",
    "pillow>=12.1.0",
    "pyautogui>=0.9.54",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "sqlalchemy>=2.0.45",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.13",
    "types-pyyaml>=6.0.12.20250915",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rezehor"]

[tool.black]
line-length = 80
target-version = ['py314']

[tool.ruff]
line-length = 80
target-version = "py314"

[tool.mypy]
python_version = "3.14"
strict = true
